#139: implement check Order#paymentState for appointed notification in AuthorizationWith3ds
