+++ functionaltests/src/test/resources/specs/paymentmethods/creditcard/AuthorizationWith3ds.html
@@ -175,6 +175,33 @@
                     <td></td>
                 </tr>
             </table>
+
+            <h3>Verify Order#paymentState</h3>
+            <p>For <b>appointed</b> payments it expected to be <b>Pending</b>,
+                for failed <i>Payment 2</i> - unchanged (null)</p>
+            <table c:execute="#orderPaymentState = fetchOrderPaymentState(#paymentName)">
+                <tr>
+                    <th c:set="#paymentName">Payment Name</th>
+                    <th c:assertEquals="#orderPaymentState">Order Payment State</th>
+                </tr>
+                <tr>
+                    <td>Payment 1</td>
+                    <td>Pending</td>
+                </tr>
+                <tr>
+                    <td>Payment 2</td>
+                    <td>(null)</td>
+                </tr>
+                <tr>
+                    <td>Payment 3</td>
+                    <td>Pending</td>
+                </tr>
+                <tr>
+                    <td>Payment 4</td>
+                    <td>Pending</td>
+                </tr>
+            </table>
+
         </div>
         <h2>Further Details</h2>
         <h3>What happens if the verification fails?</h3>
