+++ functionaltests/src/test/resources/specs/paymentmethods/creditcard/AuthorizationWithout3ds.html
@@ -194,7 +194,7 @@
 
             <h3>Verify Order#paymentState</h3>
             <p>For <b>appointed</b> payments it expected to be <b>Pending</b>,
-                for failed <i>Payment 1</i> - unchanged (null)</p>
+                for failed <i>Payment 2</i> - unchanged (null)</p>
             <table c:execute="#orderPaymentState = fetchOrderPaymentState(#paymentName)">
                 <tr>
                     <th c:set="#paymentName">Payment Name</th>
